http-api: T6326: return full warning/error output through api
When running under vyos-configd, the full output is not returned through
the api, due to the early implementation 'workaround' of writing
directly to the session tty. This is corrected to return all ambient
stdout (notably warnings) and error messages directly to the originating
stdout, which is captured by the configsession instance for any
subsequent use, including http-api.